requirement.txt
python-telegram-bot==20.3
openai==1.2.3
python-dotenv==1.0.0
